web/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
celery[redis]>=5.4.0
redis>=5.2.0
python-multipart>=0.0.12
httpx>=0.27.0
PyYAML>=6.0.1
aiofiles>=24.1.0
